:sparkles: :: SchoolLife관련 저장된 데이터가 안불러와지던 문제 해결

+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/component/ForeignLanguageComponent.kt
@@ -236,7 +236,7 @@ fun enterStudentInformation(
             portfolioUrl = profileInfo.portfolioUrl,
             contactEmail = profileInfo.contactEmail,
             formOfEmployment = workConditionInfo.formOfEmployment,
-            gsmAuthenticationScore = schoolLifeInfo.gsmAuthenticationScore,
+            gsmAuthenticationScore = schoolLifeInfo.gsmAuthenticationScore.toInt(),
             salary = workConditionInfo.salary.toInt(),
             region = workConditionInfo.region,
             languageCertificate = languageCertificate,
